moved filehandeling
