+++ src/robot/brick.rs
@@ -1,52 +1,15 @@
-use std::io::Error;
-use std::fs::File;
-use std::fs::OpenOptions;
-use std::io::{BufReader, Read, Write};
+use crate::robot::files;
 
-// functions for handeling file reads
-pub fn read_str_file(path: &str) -> String {
-    let file_result: Result<File, Error> = File::open(path);
-    let file:File = match file_result {
-        Ok(file) => file,
-        Err(error) => panic!("Faild to read file: {:?}, Error: {:?}", path, error)
-    };
-    let mut file_buff:BufReader<File> = BufReader::new(file);
+// these function will handel the ev3 brick hardware
 
-    let mut data:String = String::new();
-    file_buff.read_to_string(&mut data).expect("Faild to read string");
-    return data;
+// TODO: add functions to read buttons
+
+// functions for controlling the LED's
+pub fn clear_led() {
+    let led_paths:[&str; 4] = ["/sys/class/leds/led0:green:brick-status", "/sys/class/leds/led0:red:brick-status", "/sys/class/leds/led1:green:brick-status", "/sys/class/leds/led1:red:brick-status"];
+    for led in led_paths {
+        files::write_int(led, &0);
+    }
 }
 
-pub fn read_int_file(path: &str) -> i64 {
-    let file_result: Result<File, Error> = File::open(path);
-    let file:File = match file_result {
-        Ok(file) => file,
-        Err(error) => panic!("Faild to read file: {:?}, Error: {:?}", path, error)
-    };
-    let mut file_buff = BufReader::new(file);
-
-    let mut raw_data:String = String::new();
-    file_buff.read_to_string(&mut raw_data).expect("Faild to read int");
-
-    let data: i64 = raw_data.trim().parse::<i64>().unwrap(); // not to self: do not forget to trim the string when type casting
-    return data;
-}
-
-// function for handeling file writes
-pub fn write_str_file(path: &str, data: &str) {
-    let file_result:Result<File, Error> = OpenOptions::new().write(true).open(path);
-    let mut file:File = match file_result {
-        Ok(file) => file,
-        Err(error) => panic!("Faild to write file: {:?} data: {:?} Error: {:?}", path, data, error)
-    };
-    file.write_all(format!("{}", data).as_bytes()).unwrap();
-}
-
-pub fn write_int_file(path: &str, data: &i64) {
-    let file_result:Result<File, Error> = OpenOptions::new().write(true).open(path);
-    let mut file:File = match file_result {
-        Ok(file) => file,
-        Err(error) => panic!("Faild to write file: {:?} data: {:?} Error: {:?}", path, data, error)
-    };
-    file.write_all(format!("{}", data).as_bytes()).unwrap();
-}
+pub fn set_led(led: i64)
