+++ src/robot/mod.rs
@@ -3,3 +3,4 @@
 pub mod motor;
 pub mod brick;
 pub mod sensor;
+pub mod files;
